+++ CHANGELOG.md
@@ -9,6 +9,9 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ### Added
 - `pread` hook (used by `dotnet`);
+- Layer integration tests with more apps (testing with Go only on MacOS because of
+  known crash on Linux - [[#380](https://github.com/metalbear-co/mirrord/issues/380)]).
+  Closes [[#472](https://github.com/metalbear-co/mirrord/issues/472)].
 
 ## Changed
 - Don't report InProgress io error as error (log as info)